pyproject.toml
[tool.poetry]
name = "ecommerce-backend"
version = "0.1.0"
description = "Django REST backend for Libazkart"
authors = ["You <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^5.0"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pydantic = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"